backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
python-multipart>=0.0.6
aiohttp>=3.9.0
pillow>=10.2.0
opencv-python-headless>=4.8.0
googlemaps>=4.7.0
stripe>=7.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
bcrypt>=4.1.0
email-validator>=2.1.0
alembic>=1.12.0
slowapi>=0.1.9
pydantic[email]>=2.6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0